can now choose branch to run

+++ run.bash
@@ -1,16 +1,24 @@
 #!/bin/bash
 
 function run {
-  rm -r /srv/http/git-preview/* 2>/dev/null
+  local bran="$1"
+  [ -z "$1" ] && bran="gh-pages"
   
-  local b="$(git rev-parse --abbrev-ref HEAD)"
-  if [ "$b" == "gh-pages" ]; then
-    cp -r * /srv/http/git-preview
+  # http://stackoverflow.com/questions/5167957/is-there-a-better-way-to-find-out-if-a-local-git-branch-exists
+  if git show-ref --verify --quiet "refs/heads/$bran"; then
+    rm -r /srv/http/git-preview/* 2>/dev/null
+    
+    local b="$(git rev-parse --abbrev-ref HEAD)"
+    if [ "$b" == "$bran" ]; then
+      cp -r * /srv/http/git-preview
+    else
+      git archive "$bran" | tar -xC /srv/http/git-preview
+    fi
+    
+    git deps -l -f /srv/http/git-preview/deps -o /srv/http/git-preview
   else
-    git archive gh-pages | tar -xC /srv/http/git-preview
+    echo "Error: no branch $bran" >&2
   fi
-  
-  git deps -l -f /srv/http/git-preview/deps -o /srv/http/git-preview
 }
 
-run
+run "$@"
